+++ packages/transaction-controller/src/TransactionController.test.ts
@@ -46,6 +46,7 @@ import { MethodDataHelper } from './helpers/MethodDataHelper';
 import { MultichainTrackingHelper } from './helpers/MultichainTrackingHelper';
 import { PendingTransactionTracker } from './helpers/PendingTransactionTracker';
 import { shouldResimulate } from './helpers/ResimulateHelper';
+import { ExtraTransactionsPublishHook } from './hooks/ExtraTransactionsPublishHook';
 import type {
   AllowedActions,
   AllowedEvents,
@@ -66,6 +67,7 @@ import type {
   GasFeeFlowResponse,
   SubmitHistoryEntry,
   InternalAccount,
+  PublishHook,
 } from './types';
 import {
   GasFeeEstimateType,
@@ -104,6 +106,9 @@ type UnrestrictedMessenger = Messenger<
 
 const MOCK_V1_UUID = '9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d';
 const TRANSACTION_HASH_MOCK = '0x123456';
+const DATA_MOCK = '0x12345678';
+const VALUE_MOCK = '0xabcd';
+const ORIGIN_MOCK = 'test.com';
 
 jest.mock('@metamask/eth-query');
 jest.mock('./api/accounts-api');
@@ -115,6 +120,7 @@ jest.mock('./helpers/IncomingTransactionHelper');
 jest.mock('./helpers/MethodDataHelper');
 jest.mock('./helpers/MultichainTrackingHelper');
 jest.mock('./helpers/PendingTransactionTracker');
+jest.mock('./hooks/ExtraTransactionsPublishHook');
 jest.mock('./utils/batch');
 jest.mock('./utils/gas');
 jest.mock('./utils/gas-fees');
@@ -1610,6 +1616,7 @@ describe('TransactionController', () => {
         dappSuggestedGasFees: undefined,
         delegationAddress: undefined,
         deviceConfirmedOn: undefined,
+        disableGasBuffer: undefined,
         id: expect.any(String),
         isFirstTimeInteraction: undefined,
         nestedTransactions: undefined,
@@ -2219,6 +2226,63 @@ describe('TransactionController', () => {
         ]);
       });
 
+      it('uses extra transactions publish hook if batch transactions in metadata', async () => {
+        const { controller } = setupController({
+          messengerOptions: {
+            addTransactionApprovalRequest: {
+              state: 'approved',
+            },
+          },
+        });
+
+        const publishHook: jest.MockedFn<PublishHook> = jest.fn();
+
+        publishHook.mockResolvedValueOnce({
+          transactionHash: TRANSACTION_HASH_MOCK,
+        });
+
+        const extraTransactionsPublishHook = jest.mocked(
+          ExtraTransactionsPublishHook,
+        );
+
+        extraTransactionsPublishHook.mockReturnValue({
+          getHook: () => publishHook,
+        } as unknown as ExtraTransactionsPublishHook);
+
+        const { result, transactionMeta } = await controller.addTransaction(
+          {
+            from: ACCOUNT_MOCK,
+            to: ACCOUNT_MOCK,
+          },
+          {
+            networkClientId: NETWORK_CLIENT_ID_MOCK,
+          },
+        );
+
+        controller.updateBatchTransactions({
+          transactionId: transactionMeta.id,
+          batchTransactions: [
+            { data: DATA_MOCK, to: ACCOUNT_2_MOCK, value: VALUE_MOCK },
+          ],
+        });
+
+        result.catch(() => {
+          // Intentionally empty
+        });
+
+        await flushPromises();
+
+        expect(ExtraTransactionsPublishHook).toHaveBeenCalledTimes(1);
+        expect(ExtraTransactionsPublishHook).toHaveBeenCalledWith({
+          addTransactionBatch: expect.any(Function),
+          transactions: [
+            { data: DATA_MOCK, to: ACCOUNT_2_MOCK, value: VALUE_MOCK },
+          ],
+        });
+
+        expect(publishHook).toHaveBeenCalledTimes(1);
+      });
+
       describe('fails', () => {
         /**
          * Test template to assert adding and submitting a transaction fails.
@@ -2626,7 +2690,7 @@ describe('TransactionController', () => {
     });
 
     describe('with batch ID', () => {
-      it('throws if duplicate', async () => {
+      it('throws if duplicate and external origin', async () => {
         const { controller } = setupController({
           options: {
             state: {
@@ -2649,11 +2713,12 @@ describe('TransactionController', () => {
           controller.addTransaction(txParams, {
             batchId: BATCH_ID_MOCK,
             networkClientId: NETWORK_CLIENT_ID_MOCK,
+            origin: ORIGIN_MOCK,
           }),
         ).rejects.toThrow('Batch ID already exists');
       });
 
-      it('throws if duplicate with different case', async () => {
+      it('throws if duplicate with different case and external origin', async () => {
         const { controller } = setupController({
           options: {
             state: {
@@ -2676,8 +2741,34 @@ describe('TransactionController', () => {
           controller.addTransaction(txParams, {
             batchId: BATCH_ID_MOCK.toUpperCase() as Hex,
             networkClientId: NETWORK_CLIENT_ID_MOCK,
+            origin: ORIGIN_MOCK,
           }),
         ).rejects.toThrow('Batch ID already exists');
+      });
+
+      it('does not throw if duplicate but internal origin', async () => {
+        const { controller } = setupController({
+          options: {
+            state: {
+              transactions: [
+                {
+                  batchId: BATCH_ID_MOCK,
+                } as unknown as TransactionMeta,
+              ],
+            },
+          },
+          updateToInitialState: true,
+        });
+
+        const txParams = {
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_MOCK,
+        };
+
+        await controller.addTransaction(txParams, {
+          batchId: BATCH_ID_MOCK,
+          networkClientId: NETWORK_CLIENT_ID_MOCK,
+        });
       });
     });
   });
@@ -5245,6 +5336,42 @@ describe('TransactionController', () => {
         expect.any(String),
       );
     });
+
+    it('supports publish hook override per call', async () => {
+      const publishHookController = jest.fn();
+
+      const publishHookCall = jest.fn().mockResolvedValueOnce({
+        transactionHash: TRANSACTION_HASH_MOCK,
+      });
+
+      const { controller } = setupController({
+        options: {
+          hooks: {
+            publish: publishHookController,
+          },
+        },
+        messengerOptions: {
+          addTransactionApprovalRequest: {
+            state: 'approved',
+          },
+        },
+      });
+
+      jest.spyOn(mockEthQuery, 'sendRawTransaction');
+
+      const { result } = await controller.addTransaction(paramsMock, {
+        networkClientId: NETWORK_CLIENT_ID_MOCK,
+        publishHook: publishHookCall,
+      });
+
+      await result;
+
+      expect(controller.state.transactions[0].hash).toBe(TRANSACTION_HASH_MOCK);
+
+      expect(publishHookCall).toHaveBeenCalledTimes(1);
+      expect(publishHookController).not.toHaveBeenCalled();
+      expect(mockEthQuery.sendRawTransaction).not.toHaveBeenCalled();
+    });
   });
 
   describe('updateSecurityAlertResponse', () => {
@@ -5963,6 +6090,42 @@ describe('TransactionController', () => {
       );
 
       expect(updatedTransaction?.txParams).toStrictEqual(params);
+    });
+
+    it('updates EIP-1559 properties and returns updated transaction metadata', async () => {
+      const transactionMeta1559 = {
+        ...transactionMeta,
+        txParams: {
+          ...transactionMeta.txParams,
+          gasPrice: undefined,
+          maxFeePerGas: '0xdef',
+          maxPriorityFeePerGas: '0xabc',
+        },
+      };
+
+      const params1559: Partial<TransactionParams> = {
+        ...params,
+        maxFeePerGas: '0x456',
+        maxPriorityFeePerGas: '0x123',
+      };
+
+      delete params1559.gasPrice;
+
+      const { controller } = setupController({
+        options: {
+          state: {
+            transactions: [transactionMeta1559],
+          },
+        },
+        updateToInitialState: true,
+      });
+
+      const updatedTransaction = await controller.updateEditableParams(
+        transactionId,
+        params1559,
+      );
+
+      expect(updatedTransaction?.txParams).toStrictEqual(params1559);
     });
 
     it('updates transaction layer 1 gas fee updater', async () => {